+++ hyperbole.cabal
@@ -31,9 +31,14 @@ library
       Web.Hyperbole
       Web.Hyperbole.Application
       Web.Hyperbole.Effect
+      Web.Hyperbole.Effect.Server
       Web.Hyperbole.Embed
+      Web.Hyperbole.Event
       Web.Hyperbole.Forms
+      Web.Hyperbole.Host
       Web.Hyperbole.HyperView
+      Web.Hyperbole.Request
+      Web.Hyperbole.Response
       Web.Hyperbole.Route
       Web.Hyperbole.Session
       Web.Hyperbole.View
