+++ src/Web/Hyperbole/Host.hs
@@ -1,0 +1,7 @@
+module Web.Hyperbole.Host (Host (..)) where
+
+import Data.ByteString qualified as BS
+
+
+newtype Host = Host {text :: BS.ByteString}
+  deriving (Show)
